Update wheel env
