+++ tests/unit/test_config.py
@@ -108,10 +108,8 @@ class TestConfig:
         assert "Timeout reached. Defaulting to not enabling telemetry." in captured.out
 
     def test_generate_and_read(self, monkeypatch, tmp_path):
-        # Mock is_running_tests to return False
         monkeypatch.setattr(pybamm.config, "is_running_tests", lambda: False)
-
-        # Mock ask_user_opt_in to return True
+        monkeypatch.setattr(pybamm.config, "check_opt_out", lambda: False)
         monkeypatch.setattr(pybamm.config, "ask_user_opt_in", lambda: True)
 
         # Mock telemetry capture
